Updated Segment to make indexes
Added a function call to Segment's Update that will call a new function
in base class that will either add a new index or update the old one.
This index table can also be used to find out which vehicleid and type
and traveltype are being used on a given segment but it will have no
code implementation it is only a fast look up table.